+++ src/base/b-matryoshka/interface.ts
@@ -6,6 +6,9 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+/**
+ * Element data
+ */
 export interface Doll extends Dictionary {
 	id: string;
 	parentId?: string;